requirements.txt
google-cloud-storage==2.14.0
python-dotenv==1.0.0